Server_v1/requirements.txt
APScheduler==3.6.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-middleware==1.3.1
django-filter==2.1.0
django-jet==1.0.8
django-qrcode==0.3
django-rest-swagger==2.2.0
djangorestframework==3.9.4
google-api-python-client==1.4.1
httplib2==0.12.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
oauth2client==4.1.3
openapi-codec==1.3.2
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.3
